Module ph::fmph::goindexing

Utils for indexing with group optimization.

Structs

Bits

Size in bits.

Bits8

Seed size of 8 bits.

TwoToPowerBits

Size being the power of two.

TwoToPowerBitsStatic

Seed size given as a power of two (knowing at compile time).

Traits

GroupSize

Implementations of GroupSize represent group size in fingerprinting-based minimal perfect hashing with group optimization.

SeedSize

Implementations of SeedSize represent seed size in fingerprinting-based minimal perfect hashing with group optimization.

Functions

group_nr

Calculates group number for a given key at level of the size level_size_groups groups, whose number is already hashed in hasher. Modifies hasher, which can be farther used to calculate index in the group by just writing to it the seed of the group.