Enum ph::fmph::keyset::CachedKeySet

pub enum CachedKeySet<K, KS> {
    Dynamic(KS, usize),
    Cached(Vec<K>),
}

Implementation of KeySet that initially stores another KeySet (which is usually succinct but slow, such as DynamicKeySet), but when number of keys drops below given threshold, the remaining keys are cached (cloned into the vector), and later only the cache is used.

Variants

Dynamic(KS, usize)

Cached(Vec<K>)

Implementations

impl<K, KS> CachedKeySet<K, KS>

pub fn new(key_set: KS, clone_threshold: usize) -> Self

Constructs cached key_set. The keys are cloned and cached as soon as their number drops below clone_threshold.

impl<K, PI: GetIterator> CachedKeySet<K, DynamicKeySet<PI>>

pub fn dynamic(keys: PI, clone_threshold: usize) -> Self

Constructs cached DynamicKeySet that obtains the keys by keys that returns iterator over keys. The keys are cloned and cached as soon as their number drops below clone_threshold.

If the number of keys is known, it is more efficient to call CachedKeySet::dynamic_with_len instead.

Example

use {ph::fmph::keyset::{KeySet, CachedKeySet}, rayon::iter::{ParallelIterator, IntoParallelIterator}};
// Constructing a dynamic key set consisting of the squares of the integers from 1 to 100,
// part of which will be cached by the first call of any of the retain methods:
let ks = CachedKeySet::dynamic(|| (1..=100).map(|v| v*v), usize::MAX);
assert_eq!(ks.keys_len(), 100);
// Same as above but using multiple threads to generate keys:
let ks = CachedKeySet::dynamic((|| (1..=100).map(|v| v*v), || (1..=100).into_par_iter().map(|v| v*v)), usize::MAX);
assert_eq!(ks.keys_len(), 100);

pub fn dynamic_with_len(keys: PI, len: usize, clone_threshold: usize) -> Self

Constructs cached DynamicKeySet that obtains the keys by keys that returns iterator over exactly len keys. The keys are cloned and cached as soon as their number drops below clone_threshold.

Example

use {ph::fmph::keyset::{KeySet, CachedKeySet}, rayon::iter::{ParallelIterator, IntoParallelIterator}};
// Constructing a dynamic key set consisting of the squares of the integers from 1 to 100,
// part of which will be cached by the first call of any of the retain methods:
let ks = CachedKeySet::dynamic_with_len(|| (1..=100).map(|v| v*v), 100, usize::MAX);
assert_eq!(ks.keys_len(), 100);
// Same as above but using multiple threads to generate keys:
let ks = CachedKeySet::dynamic_with_len((|| (1..=100).map(|v| v*v), || (1..=100).into_par_iter().map(|v| v*v)), 100, usize::MAX);
assert_eq!(ks.keys_len(), 100);

impl<'k, K: Sync> CachedKeySet<K, SliceSourceWithRefs<'k, K>>

pub fn slice(keys: &'k [K], clone_threshold: usize) -> Self

Constructs cached SliceSourceWithRefs that wraps given keys. The keys are cloned and cached as soon as their number drops below clone_threshold. After cloning, the keys are placed in a continuous memory area which is friendly to the CPU cache.