Struct ph::fmph::keyset::DynamicKeySet

pub struct DynamicKeySet<GetKeyIter: GetIterator> {
    pub keys: GetKeyIter,
    pub len: usize,
}

Implementation of KeySet that stores only the object that returns iterator over all keys (the iterator can even expose the keys that have been removed earlier by retain methods). It is usually a good idea to use it within CachedKeySet, see CachedKeySet::dynamic.

Fields
keys: GetKeyIterlen: usize
Implementations

impl<GetKeyIter: GetIterator> DynamicKeySet<GetKeyIter>

pub fn new(keys: GetKeyIter) -> Self

Constructs a DynamicKeySet that obtains the keys by keys.

pub fn with_len(keys: GetKeyIter, len: usize) -> Self

Constructs a DynamicKeySet that obtains the keys by keys (which should produce len keys).