Struct ph::fmph::BuildConf

pub struct BuildConf<S = BuildDefaultSeededHasher> {
    pub hash_builder: S,
    pub cache_threshold: usize,
    pub relative_level_size: u16,
    pub use_multiple_threads: bool,
}

Build configuration that is accepted by Function constructors.

See field descriptions for details.

Fields
hash_builder: S

The family of hash functions used by the constructed FMPH. (default: BuildDefaultSeededHasher)

cache_threshold: usize

The threshold for the number of keys below which their hashes will be cached during level construction. (default: BuildConf::DEFAULT_CACHE_THRESHOLD)

Caching speeds up level construction at the expense of memory consumption during construction (caching a single key requires 8 bytes of memory). Caching is particularly recommended for keys with complex types whose hashing is slow. It is possible to use a value of 0 to disable caching completely, or usize::MAX to use it on all levels.

relative_level_size: u16

Size of each level given as a percentage of the number of level input keys. (default: 100)

A value of 100 minimizes the size of the constructed minimum perfect hash function. Larger values speed up evaluation at the expense of increased size. For example, the values 100 and 200 lead to the sizes of approximately 2.8 and 3.4 bits per input key, respectively. It does not make sense to use values below 100.

use_multiple_threads: bool

Whether to use multiple threads during construction. (default: true)

If true, the construction will be performed using the default rayon thread pool.

Implementations

impl BuildConf

pub fn mt(use_multiple_threads: bool) -> Self

Returns configuration that potentially uses multiple threads to build Function.

pub fn ct(cache_threshold: usize) -> Self

Returns configuration that uses custom cache_threshold to build Function.

pub fn ct_mt(cache_threshold: usize, use_multiple_threads: bool) -> Self

Returns configuration that uses custom cache_threshold and potentially uses multiple threads to build Function.

pub fn lsize(relative_level_size: u16) -> Self

Returns configuration that uses at each level a bit-array of size relative_level_size given as a percent of number of level input keys.

pub fn lsize_ct(relative_level_size: u16, cache_threshold: usize) -> Self

Returns configuration that uses custom relative_level_size and cache_threshold to build Function.

pub fn lsize_mt(relative_level_size: u16, use_multiple_threads: bool) -> Self

Returns configuration that potentially uses multiple threads and at each level a bit-array of size relative_level_size given as a percent of number of level input keys.

impl<S> BuildConf<S>

pub const DEFAULT_CACHE_THRESHOLD: usize = 134_217_728usize

The default value for relative_level_size, which results in building the cache with a maximum size of 1GB.

pub fn hash(hash_builder: S) -> Self

Returns configuration that uses custom hash_builder.

pub fn hash_lsize(hash_builder: S, relative_level_size: u16) -> Self

Returns configuration that uses custom hash_builder and relative_level_size.

pub fn hash_lsize_mt( hash_builder: S, relative_level_size: u16, use_multiple_threads: bool ) -> Self

Returns configuration that uses custom hash_builder, relative_level_size and potentially uses multiple threads to build Function.

pub fn hash_lsize_ct( hash_builder: S, relative_level_size: u16, cache_threshold: usize ) -> Self

Returns configuration that uses custom hash_builder, relative_level_size and cache_threshold to build Function.

pub fn hash_lsize_ct_mt( hash_builder: S, relative_level_size: u16, cache_threshold: usize, use_multiple_threads: bool ) -> Self

Returns configuration that uses custom hash_builder, relative_level_size, cache_threshold and potentially uses multiple threads to build Function.